Added SKIP_BUNDLE_CACHE var to .expeditor/buildkite/verify.sh
This lets me test our docker setup locally much quicker as I don't
have to time out on the aws s3 push/pull steps.

+++ .expeditor/buildkite/verify.sh
@@ -17,7 +17,11 @@ push_s3_file() {
 }
 
 apt-get update -y
-apt-get install sudo awscli -y
+apt-get install sudo -y
+
+if [ -z "${SKIP_BUNDLE_CACHE:-}" ]; then
+    apt-get install awscli -y
+fi
 
 echo "--- setting up user"
 useradd -m -U --uid 2000 normal
@@ -32,32 +36,36 @@ gem env
 bundle --version
 
 echo "--- bundle install"
-pull_s3_file "bundle.tar.gz"
-pull_s3_file "bundle.sha256"
+if [ -z "${SKIP_BUNDLE_CACHE:-}" ]; then
+    pull_s3_file "bundle.tar.gz"
+    pull_s3_file "bundle.sha256"
 
-if [ -f bundle.tar.gz ]; then
-  tar -xzf bundle.tar.gz
-fi
+    if [ -f bundle.tar.gz ]; then
+        tar -xzf bundle.tar.gz
+    fi
 
-if [ -n "${RESET_BUNDLE_CACHE:-}" ]; then
-    rm bundle.sha256
+    if [ -n "${RESET_BUNDLE_CACHE:-}" ]; then
+        rm bundle.sha256
+    fi
 fi
 
 bundle config --local path vendor/bundle
 bundle install --jobs=7 --retry=3 --without tools maintenance deploy
 
 echo "--- bundle cache"
-if test -f bundle.sha256 && shasum --check bundle.sha256 --status; then
-    echo "Bundled gems have not changed. Skipping upload to s3"
-else
-    echo "Bundled gems have changed. Uploading to s3"
-    shasum -a 256 Gemfile.lock > bundle.sha256
-    tar -czf bundle.tar.gz vendor/
-    push_s3_file bundle.tar.gz
-    push_s3_file bundle.sha256
+if [ -z "${SKIP_BUNDLE_CACHE:-}" ]; then
+    if test -f bundle.sha256 && shasum --check bundle.sha256 --status; then
+        echo "Bundled gems have not changed. Skipping upload to s3"
+    else
+        echo "Bundled gems have changed. Uploading to s3"
+        shasum -a 256 Gemfile.lock > bundle.sha256
+        tar -czf bundle.tar.gz vendor/
+        push_s3_file bundle.tar.gz
+        push_s3_file bundle.sha256
+    fi
 fi
 
-echo "+++ bundle exec rake"
+echo "+++ bundle exec rake test"
 # TODO: run this as non-root:
 # chown -R normal:normal /home/normal /workdir
 # su normal -c "bundle exec rake ${RAKE_TASK:-}"
